framework for 0d parcel test
